Tests complited

+++ src/main/java/com/schedule/proj/repository/TeacherRepository.java
@@ -14,5 +14,5 @@ import java.util.Set;
 
 @Repository
 public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
-    Teacher getByUserId(int id);
+    Teacher getByUserId(Long id);
 }
