+++ src/main/java/com/schedule/proj/service/TeacherService.java
@@ -74,7 +74,7 @@ public class TeacherService {
         String email = jwtProvider.getLoginFromToken(token);
         User user = userRepository.findUserByEmail(email);
         if(user.getUserRole() == UserRole.TEACHER && user!=null) {
-            Teacher teacher = teacherRepository.getByUserId(user.getId());
+            Teacher teacher = teacherRepository.getByUserId((long) user.getId());
             user.setFirstName(dto.getFirstname());
             user.setLastName(dto.getLastname());
             teacher.setCathedra(dto.getCathedra());
